+++ client/src/pages/Signup/Signup.js
@@ -1,42 +1,22 @@
-import React, { useState } from "react";
+import React from "react";
 import { Redirect, useHistory } from "react-router-dom";
 import { connect } from "react-redux";
 import {
   Grid,
-  Box,
   Typography,
-  Button,
-  FormControl,
-  TextField,
-  FormHelperText,
-  Container,
-  SvgIcon
+  Box,
+  Button
 } from "@material-ui/core";
+
+import SignupForm from "./SignupForm";
 import { register } from "../../store/utils/thunkCreators";
-import background from "../../assets/bg-img.png";
 import { ReactComponent as BubbleIcon } from '../../assets/bubble.svg';
 
-import './Signup.css'
+import './Signup.css';
 
-const Login = (props) => {
-  const history = useHistory();
+const Signup = (props) => {
   const { user, register } = props;
-  const [formErrorMessage, setFormErrorMessage] = useState({});
-
-  const handleRegister = async (event) => {
-    event.preventDefault();
-    const username = event.target.username.value;
-    const email = event.target.email.value;
-    const password = event.target.password.value;
-    const confirmPassword = event.target.confirmPassword.value;
-
-    if (password !== confirmPassword) {
-      setFormErrorMessage({ confirmPassword: "Passwords must match" });
-      return;
-    }
-
-    await register({ username, email, password });
-  };
+  const history = useHistory();
 
   if (user.id) {
     return <Redirect to="/home" />;
@@ -73,93 +53,9 @@ const Login = (props) => {
               onClick={() => history.push("/login")}
             >
               Login
-          </Button>
+            </Button>
           </Grid>
-          <form onSubmit={handleRegister}>
-            <Grid>
-              <Typography variant="h5">Create an account.</Typography>
-              <Grid>
-                <FormControl
-                  fullWidth
-                  margin="normal"
-                >
-                  <TextField
-                    aria-label="username"
-                    label="Username"
-                    name="username"
-                    type="text"
-                    required
-                  />
-                </FormControl>
-              </Grid>
-              <Grid>
-                <FormControl
-                  fullWidth
-                  margin="normal"
-                >
-                  <TextField
-                    label="E-mail address"
-                    aria-label="e-mail address"
-                    type="email"
-                    name="email"
-                    required
-                  />
-                </FormControl>
-              </Grid>
-              <Grid>
-                <FormControl
-                  fullWidth
-                  margin="normal"
-                  error={!!formErrorMessage.confirmPassword}
-                >
-                  <TextField
-                    aria-label="password"
-                    label="Password"
-                    type="password"
-                    inputProps={{ minLength: 6 }}
-                    name="password"
-                    required
-                  />
-                  <FormHelperText>
-                    {formErrorMessage.confirmPassword}
-                  </FormHelperText>
-                </FormControl>
-              </Grid>
-              <Grid>
-                <FormControl
-                  fullWidth
-                  margin="normal"
-                  error={!!formErrorMessage.confirmPassword}
-                >
-                  <TextField
-                    label="Confirm Password"
-                    aria-label="confirm password"
-                    type="password"
-                    inputProps={{ minLength: 6 }}
-                    name="confirmPassword"
-                    required
-                  />
-                  <FormHelperText>
-                    {formErrorMessage.confirmPassword}
-                  </FormHelperText>
-                </FormControl>
-              </Grid>
-              <Grid
-                container
-                item
-                justify="center"
-              >
-                <Button
-                  type="submit"
-                  variant="contained"
-                  size="large"
-                  color="primary"
-                >
-                  Create
-              </Button>
-              </Grid>
-            </Grid>
-          </form>
+          <SignupForm register={register} />
         </Box>
       </Grid>
     </Grid>
@@ -180,4 +76,4 @@ const mapDispatchToProps = (dispatch) => {
   };
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(Login);
+export default connect(mapStateToProps, mapDispatchToProps)(Signup);
